+++ benches/bench.rs
@@ -16,14 +16,14 @@ fn bench_3_smooth(c: &mut Criterion) {
         group.bench_with_input(BenchmarkId::new("pratt", n), &n, |b, &n| {
             b.iter(|| pratt(n))
         });
+        group.bench_with_input(BenchmarkId::new("smooth", n), &(2, n), |b, &(k, n)| {
+            b.iter(|| smooth(k, n))
+        });
         group.bench_with_input(
             BenchmarkId::new("with_primes", n),
             &([2, 3], &n),
             |b, (primes, &n)| b.iter(|| with_primes(primes, n)),
         );
-        // group.bench_with_input(BenchmarkId::new("smooth", n), &(2, n), |b, &(k, n)| {
-        //     b.iter(|| smooth(k, n))
-        // });
     }
 
     group.finish();